+++ src/lib.rs
@@ -9,4 +9,4 @@ mod chat;
 mod jobs;
 mod routes;
 pub mod settings;
-pub mod startup;
+pub mod start;
