+++ src/main.rs
@@ -4,7 +4,7 @@
 use telebot::settings::environment::get_environment;
 use telebot::settings::get_settings;
 use telebot::settings::telemetry::init_tracing;
-use telebot::startup::start_app;
+use telebot::start::start_app;
 
 #[tokio::main]
 async fn main() {
